requirements.txt
asgiref==3.8.1
black==25.1.0
click==8.1.8
contourpy==1.3.1
crispy==0.8.0
crispy-bootstrap5==2024.10
cycler==0.12.1
Django==5.1.6
django-crispy-forms==2.3
django-debug-toolbar==5.0.1
django-postman==4.5
django-widget-tweaks==1.5.0
fabio==2024.9.0
fonttools==4.56.0
h5py==3.12.1
hdf5plugin==5.0.0
kiwisolver==1.4.8
lxml==5.3.1
matplotlib==3.10.0
mypy-extensions==1.0.0
numpy==2.2.3
packaging==24.2
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
psycopg2-binary==2.9.10
pyparsing==3.2.1
PyQt6==6.8.1
PyQt6-Qt6==6.8.2
PyQt6_sip==13.10.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
ruamel.yaml==0.18.10
scipy==1.15.2
silx==2.2.0
six==1.17.0
SQLAlchemy==2.0.38
sqlparse==0.5.3
typing_extensions==4.12.2
whitenoise==6.9.0
xraydb==4.5.6
